+++ TestApplications/src/MFCApplication/MFCApplicationView.cpp
@@ -1,0 +1,128 @@
+
+// MFCApplicationView.cpp : implementation of the CMFCApplicationView class
+//
+
+#include "pch.h"
+#include "framework.h"
+// SHARED_HANDLERS can be defined in an ATL project implementing preview, thumbnail
+// and search filter handlers and allows sharing of document code with that project.
+#ifndef SHARED_HANDLERS
+#include "MFCApplication.h"
+#endif
+
+#include "MFCApplicationDoc.h"
+#include "MFCApplicationView.h"
+
+#ifdef _DEBUG
+#define new DEBUG_NEW
+#endif
+
+
+// CMFCApplicationView
+
+IMPLEMENT_DYNCREATE(CMFCApplicationView, CView)
+
+BEGIN_MESSAGE_MAP(CMFCApplicationView, CView)
+	// Standard printing commands
+	ON_COMMAND(ID_FILE_PRINT, &CView::OnFilePrint)
+	ON_COMMAND(ID_FILE_PRINT_DIRECT, &CView::OnFilePrint)
+	ON_COMMAND(ID_FILE_PRINT_PREVIEW, &CMFCApplicationView::OnFilePrintPreview)
+	ON_WM_CONTEXTMENU()
+	ON_WM_RBUTTONUP()
+END_MESSAGE_MAP()
+
+// CMFCApplicationView construction/destruction
+
+CMFCApplicationView::CMFCApplicationView() noexcept
+{
+	// TODO: add construction code here
+
+}
+
+CMFCApplicationView::~CMFCApplicationView()
+{
+}
+
+BOOL CMFCApplicationView::PreCreateWindow(CREATESTRUCT& cs)
+{
+	// TODO: Modify the Window class or styles here by modifying
+	//  the CREATESTRUCT cs
+
+	return CView::PreCreateWindow(cs);
+}
+
+// CMFCApplicationView drawing
+
+void CMFCApplicationView::OnDraw(CDC* /*pDC*/)
+{
+	CMFCApplicationDoc* pDoc = GetDocument();
+	ASSERT_VALID(pDoc);
+	if (!pDoc)
+		return;
+
+	// TODO: add draw code for native data here
+}
+
+
+// CMFCApplicationView printing
+
+
+void CMFCApplicationView::OnFilePrintPreview()
+{
+#ifndef SHARED_HANDLERS
+	AFXPrintPreview(this);
+#endif
+}
+
+BOOL CMFCApplicationView::OnPreparePrinting(CPrintInfo* pInfo)
+{
+	// default preparation
+	return DoPreparePrinting(pInfo);
+}
+
+void CMFCApplicationView::OnBeginPrinting(CDC* /*pDC*/, CPrintInfo* /*pInfo*/)
+{
+	// TODO: add extra initialization before printing
+}
+
+void CMFCApplicationView::OnEndPrinting(CDC* /*pDC*/, CPrintInfo* /*pInfo*/)
+{
+	// TODO: add cleanup after printing
+}
+
+void CMFCApplicationView::OnRButtonUp(UINT /* nFlags */, CPoint point)
+{
+	ClientToScreen(&point);
+	OnContextMenu(this, point);
+}
+
+void CMFCApplicationView::OnContextMenu(CWnd* /* pWnd */, CPoint point)
+{
+#ifndef SHARED_HANDLERS
+	theApp.GetContextMenuManager()->ShowPopupMenu(IDR_POPUP_EDIT, point.x, point.y, this, TRUE);
+#endif
+}
+
+
+// CMFCApplicationView diagnostics
+
+#ifdef _DEBUG
+void CMFCApplicationView::AssertValid() const
+{
+	CView::AssertValid();
+}
+
+void CMFCApplicationView::Dump(CDumpContext& dc) const
+{
+	CView::Dump(dc);
+}
+
+CMFCApplicationDoc* CMFCApplicationView::GetDocument() const // non-debug version is inline
+{
+	ASSERT(m_pDocument->IsKindOf(RUNTIME_CLASS(CMFCApplicationDoc)));
+	return (CMFCApplicationDoc*)m_pDocument;
+}
+#endif //_DEBUG
+
+
+// CMFCApplicationView message handlers