+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 
 ## [Unreleased][]
 
+## [Release][3.0.2] [3.0.2][3.0.1-3.0.2] - 2024-03-08
+
 ### Added
 
   - [#155](https://github.com/GDATASoftwareAG/robotframework-flaui/issues/155) New Keywords
@@ -592,7 +594,11 @@ General bugfixing from keywords and improvements from artifact building and test
 
 [github.io]: https://gdatasoftwareag.github.io/robotframework-flaui
 
-[unreleased]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/3.0.0...main
+[unreleased]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/3.0.2...main
+
+[3.0.2]: https://github.com/GDATASoftwareAG/robotframework-flaui/releases/tag/3.0.2
+
+[3.0.1-3.0.2]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/3.0.1...3.0.2
 
 [3.0.1]: https://github.com/GDATASoftwareAG/robotframework-flaui/releases/tag/3.0.1